Add administrative dashboard view
First iteration of a dashboard for authors and administrators. Implements
a paginated view (`/admin/dashboard/<id>`) to inspect posts from a
administrative viewpoint. As such, it shows all available information and
options (currently, only editing a post). Moreover, users and categories
can be inspected and managed as well.

Parts of the '/admin/' endpoints are only implemented as stubs, users and
categories can neither be created nor updated at the moment.