Add theme support for templates
Templates are now rendered based on a 'theme'. Themes populate the
`blog/templates/` directory as subdirectories. The active theme is
set via a configuration key named `BLOG_THEME`. To render templates
based on the active theme there is a function called `render_theme_template`
in `blog.utils` which is called exactly the same way as `render_template`
but does not need the theme subdirectory name as part of the template path.