Beautify "default" theme
This commit introduces actual content for the "default" theme. It is
heavily based on html5up.net's "future imperfect"[1] which is licensed
under CC-BY 3.0[2]. The new README mentions this fact (to comply with
CC-BY 3.0 and encourage users to make that information available as well),
same for the CSS file (blog.css).

The theme itself strifes to be bright and simple yet elegant.

[1](http://html5up.net/future-imperfect)
[2](http://html5up.net/license)